Add files via upload
In this project, I use classification methods to solve a real historical
data-mining problem, locating displaced persons living in makeshift shelters following the destruction of
the earthquake in Haiti in 2010.
Following that earthquake, rescue workers, mostly from the United States military, needed to get food
and water to the displaced persons. But with destroyed communications, impassable roads, and
thousands of square miles, locating the people was critical.
As part of the rescue effort, a team from the Rochester Institute of Technology were flying an aircraft to
collect high resolution geo-referenced imagery. It was known that the people whose homes had been
destroyed by the earthquake were creating temporary shelters using blue tarps, and these blue tarps
would be good indicators for locating the people—if only they could be located in time, out of the
thousands of images that would be collected every day. The problem was that there was no way for
people to search the thousands of images in time to find the tarps and communicate the locations back
to the rescue workers in time. The solution would be provided by data-mining algorithms, which could
search the images far faster and more thoroughly than humanly possible. The goal was to find the best
algorithm that could search the images and locate displaced persons in time for the locations to be
communicated back to the rescue workers.